+++ config/test.exs
@@ -16,8 +16,8 @@ config :td_lm, TdLm.Repo,
   pool: Ecto.Adapters.SQL.Sandbox,
   pool_size: 1
 
-config :td_lm, :audit_service, api_service: TdLmWeb.ApiServices.MockTdAuditService
 config :td_lm, permission_resolver: TdLm.Permissions.MockPermissionResolver
 config :td_lm, :relation_loader, load_on_startup: false
 
+config :td_cache, :audit, stream: "audit:events:test"
 config :td_cache, redis_host: "redis"